Reorders triggered order checks before NFT lease
Fixes issue where, with multiple NFTs, it was leasing them and
round robining through them even if later detected that the order was
already triggered. We can safely check trigger before, it's just known
open trades that could still be modified async on us and cause us to
burn the lease, but this is rare.