Fixes overlapping trading variable fetches
There are scenarios that could have triggered multiple, overlapping
fetching of trading variables and the logic that existed previously was
insufficient to prevent this. Considering how costly the act of
refreshing is RPC wise, we want to try to avoid this by debouncing
subsequent calls when there is already an outstanding fetch in progress.