Supports concurrent price update processing
Previously a flag was used to gate processing of price updates
to deal with the fact that the logic would potentially
step all over itself due to `async` calls that could happen in the
processing loop. I've patched this up by immediately adding the
triggered transaction id to the `triggeredOrders` `Map` before
proceeding with any async calls and adding logic to handle
conditions where the transaction is not sent that removes
transaction id from the `triggeredOrders` `Map` immediately.